Adds an algorithm I implemented a while ago to calculate pi.
The algorithm involved iteratively calculates the length of a line segment riding
the circumference of the top-right quadrant of a unit circle. 17 iterations is good
(with the built-in math module) to get a ten-digits precision value calculation of
pi, which is 3.1415926535, which is sufficient for my usage (for now).

The original code I wrote used mpmath, and my tests ran with 10,000 iterations, but
the built-in math module generates an OverflowError on math.pow(2, 10000), so I may
reintroduce mpmath in the future.

I added some Python-related file/folder entries to the ignore file.